refactor(web): decompose app.py into focused sibling modules
app.py was ~1006 lines mixing four concerns: the create_app factory, ~475
lines of route declarations (_register_routes, carrying a # noqa: C901,
PLR0915 waiver), the ~220-line ADR-0016 enrichment feature, and the
humanize_age Jinja filter. This extracts each into a web/ sibling module,
following the existing register_*(app, ...) pattern (web/brief.py,
web/_deps.py). Behavior-preserving — no route, response, or signature
changes.

New modules:
- web/filters.py        — humanize_age + register_filters() (wires
                          humanize_age + the ordinal filter from #108)
- web/enrichment.py     — the whole ADR-0016 block; _register_enrichment_routes
                          renamed to public register_enrichment_routes
- web/_helpers.py       — resolve_top_bills (landing + bills, cross-module)
- web/routes_search.py  — landing (/) + /search, register(app, limiter)
- web/routes_bills.py, routes_members.py, routes_votes.py,
  routes_proceedings.py — one register(app) per entity

web/_deps.py gains the vec-loaded get_db dependency (the one connection
shared across every entity route). Single-route constants travel with
their owning module; only get_db and resolve_top_bills were genuinely
cross-module — keeping them out of app.py avoids the app<->routes import
cycle (the same reason _deps.py exists).

app.py is now a 171-line assembly root: create_app() bootstraps the schema
(ADR 0012), builds app.state, registers filters, and calls each register_*.
The PLR0915 waiver is gone; the residual C901 is only on the wrappers that
declare the inherently-branchy search/enrichment handlers (which already
carried it). __all__ shrinks to ["create_app"].

Tests: only test_web_bills.py changes — humanize_age now imports from
web/filters, and the _enrich_one_bill reference points at web/enrichment.
Full suite green (700), ruff + mypy --strict clean.

Plan: docs/plans/decompose-web-app.md. Follow-up to #108.
